P0-3: make engine_params versions immutable, reproducible snapshots
Stored engine_params rows were partial (each migration wrote only the keys
it changed) and the admin reader re-ran them through the zod schema, which
back-fills missing keys with today's defaults. So a historical version read
back with current values — v2/v5/v6 all resolved identically and diffed
empty, while v1 (pre-schema) failed validation. Genuine reproducibility
defect.

- migration 20260617000004: add schema_version / params_hash / code_sha /
  is_replayable to engine_params; backfill provenance for v1-v6. Only v6
  (the active, already-complete snapshot) is replayable; v1-v5 were stored
  partial and are flagged non-replayable (reproducible as context, not
  byte-for-byte). v1 is schema_version 1 (predates the schema).
- params-provenance.ts: canonical (sorted-key) JSON hashing, resolveProvenance
  (replayable only when stored bytes deep-equal the parsed-with-defaults
  result), materializeParams. Hashing kept out of the pure engine dir.
- getEngineParamsVersion stops masking: returns the stored bytes plus a
  resolved full set, provenance, and a hash-verified flag — no silent
  materialization. propose_engine_params now stores a fully-materialized
  snapshot with hash + schema_version + code_sha + is_replayable.
- get_engine_params / replay surface provenance; diff compares stored bytes
  so legacy versions show their real deltas. replay uses the resolved set and
  refuses candidates that no longer validate.

Also fixes a latent bug in the (never-applied) metric-truth view migration:
sum() widened planned_sets/logged_sets to numeric, which create-or-replace
rejects — cast back to bigint. Both 20260617000003 and 000004 applied to the
live project. typecheck/lint/test green (238 tests).

+++ supabase/migrations/20260617000003_metric_truth_view_fixes.sql
@@ -138,8 +138,10 @@ select
   mc.is_deload,
   wes.muscle_group_id,
   mg.name as muscle_group,
-  sum(wes.prescribed_sets) as planned_sets,
-  sum(wes.logged_sets) as logged_sets
+  -- cast back to bigint so `create or replace view` keeps the original column
+  -- types (sum() of the per-we counts widens to numeric otherwise)
+  sum(wes.prescribed_sets)::bigint as planned_sets,
+  sum(wes.logged_sets)::bigint as logged_sets
 from public.microcycles mc
   join public.workouts w on w.microcycle_id = mc.id
   join we_sets wes on wes.workout_id = w.id
